Add react context api and moved all state into provider
