+++ app/src/components/App.js
@@ -1,22 +1,16 @@
-import React, { Component } from 'react';
-import TopMenu from './TopMenu';
-import MainMenu from './MainMenu';
-import Body from './Body';
+import React from 'react';
 
-class App extends Component {
-  state = {};
+import TopMenu from './TopMenu/TopMenu';
+import MainMenu from './MainMenu/MainMenu';
+import Body from './Body/Body';
+import { MyProvider } from '../provider/MyProvider';
 
-  render() {
-    const { svgCode } = this.props;
-
-    return (
-      <div>
-        <TopMenu />
-        <MainMenu />
-        <Body svgCode={svgCode} />
-      </div>
-    );
-  }
-}
+const App = ({ svgCode }) => (
+  <MyProvider>
+    <TopMenu />
+    <MainMenu />
+    <Body svgCode={svgCode} />
+  </MyProvider>
+);
 
 export default App;
