+++ app/src/components/Body/PrimaryDisplay/Diagram/Legend.js
